Auth, Prisma init done and BugFixing
