+++ app/(portal)/[folder]/[file]/page.tsx
