+++ app/components/AuthButtons.tsx
@@ -1,0 +1,10 @@
+'use server';
+import {auth, signOut} from "@/lib/auth";
+
+
+export const signOutAction = async () => {
+    'use server';
+    if ((await auth())!.user) {
+        await signOut();
+    }
+}
