+++ lib/auth.ts
@@ -1,0 +1,4 @@
+import NextAuth from "next-auth"
+import {authConfig} from "@/auth.config"
+
+export const {auth, signIn, signOut, handlers: {GET, POST}} = NextAuth(authConfig)
